bliss, sprites, lord, stasis

+++ models/spells/mass_stasis.py
@@ -13,19 +13,19 @@ class MassStasis(Spell): # mordred???????????????????????????????
     def on_init(self):
         self.name = "Mass Stasis"
         self.range = 7 
-        self.tags = [Ethereal, Tags.Enchantment]
+        self.tags = [Ethereal, Tags.Enchantment, Tags.Translocation]
         self.level = 3
 
         self.duration = 4
+        self.radius = 5
         self.damage_type = Ethereal
 
         self.max_charges = 4
-        self.can_target_empty = False
 
-        self.upgrades['max_charges'] = (3, 1)
-        self.upgrades['duration'] = (3, 1)
-        self.upgrades['radius'] = (3, 1)
-        self.upgrades['hazards'] = (3, 2, "Pertrusion", "Leave behind 3 pure ether hazards for 3 turns that deal [7_äthereal:äthereal] damage in a 5 tile radius around the rift")
+        self.upgrades['max_charges'] = (3, 2)
+        self.upgrades['duration'] = (3, 2)
+        self.upgrades['radius'] = (2, 2)
+        self.upgrades['linger'] = (1, 2, "Äther Lag", "Targets also get ätherealiesed for the same duration after they return")
 
     def can_cast(self, x, y):
         unit = self.caster.level.get_unit_at(x,y)
@@ -36,7 +36,10 @@ class MassStasis(Spell): # mordred???????????????????????????????
     def cast(self, x, y):
         for tile in self.caster.level.get_points_in_ball(x,y,self.get_stat('radius')):
             unit = self.caster.level.get_unit_at(tile.x, tile.y)
-
+            if not unit or unit == self.caster:
+                continue
+            if self.get_stat('linger'):
+                unit.apply_buff(EtherealnessBuff(),self.get_stat('duration')*2)
             if not unit.gets_clarity:
                 unit_copy = copy.copy(unit)
                 unit.kill(None, False) 
@@ -50,7 +53,7 @@ class MassStasis(Spell): # mordred???????????????????????????????
                 stasis = Stasis(self.caster,self,self.get_stat('duration'),Goblin())
                 self.caster.level.add_obj(stasis, unit.x, unit.y)
 
-            yield
+        yield
 
     def get_description(self):
-        return "Transport target into the Ether Plane for [{duration}_turns:duration].\nLeaves behind a Rift that [ätherealieses:äthereal] units for 2 turns.".format(**self.fmt_dict())
+        return "Transport all targets in [{radius}_tile:radius] radius into the Ether Plane for [{duration}_turns:duration].\nLeaves behind a Rift that [ätherealieses:äthereal] units for 2 turns.".format(**self.fmt_dict())
